Add bank method added in service, global exception handler for better communicates added

+++ src/main/java/com/remitly/axan18/swift_api/services/BankServiceImpl.java
@@ -1,10 +1,17 @@
 package com.remitly.axan18.swift_api.services;
 
+import com.remitly.axan18.swift_api.exceptions.CannotDeleteEntityException;
+import com.remitly.axan18.swift_api.mappers.BankMapper;
 import com.remitly.axan18.swift_api.models.BankDTO;
 import com.remitly.axan18.swift_api.models.BankHeadquarterDTO;
 import com.remitly.axan18.swift_api.repositories.BankRepository;
+import jakarta.persistence.EntityNotFoundException;
+import jakarta.validation.ConstraintViolationException;
+import jakarta.validation.ValidationException;
 import lombok.AllArgsConstructor;
 import org.springframework.dao.DataAccessException;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.DuplicateKeyException;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
@@ -14,22 +21,39 @@ import java.util.List;
 @AllArgsConstructor
 public class BankServiceImpl implements BankService{
     private final BankRepository bankRepository;
+    private final BankMapper bankMapper;
 
     @Override
-    public String deleteBySwift(String swiftCode){
-        try{
+    public boolean deleteBySwift(String swiftCode) {
+        if (swiftCode == null)
+            throw new IllegalArgumentException("Provided swift is null.");
+
+        if (!bankRepository.existsById(swiftCode))
+            throw new EntityNotFoundException("Bank with swift " + swiftCode + " does not exist.");
+
+        try {
             bankRepository.deleteById(swiftCode);
-            return "Bank deleted successfully";
-        }catch (EmptyResultDataAccessException e) {
-            return "Bank with swift " + swiftCode + " does not exist.";
-        } catch (IllegalArgumentException e) {
-            return "Provided swift is null.";
+            return true;
         } catch (DataAccessException e) {
-            return "Database error occurred while deleting bank: " + e.getMessage();
+            throw new CannotDeleteEntityException("Database error occurred while deleting bank", e);
         }
     }
 
-
+    @Override
+    public boolean addBank(BankDTO bankDTO) {
+        try {
+            bankRepository.save(bankMapper.toBank(bankDTO));
+            return true;
+        } catch (DataIntegrityViolationException e) {
+            throw new DuplicateKeyException("Bank with this swiftCode already exists", e);
+        } catch (ConstraintViolationException e) {
+            throw new ValidationException("Invalid bank data: " + e.getMessage(), e);
+        } catch (IllegalArgumentException e) {
+            throw new IllegalStateException("Bank data cannot be null", e);
+        } catch (Exception e) {
+            throw new RuntimeException("Unexpected error while adding bank", e);
+        }
+    }
 
     @Override
     public List<BankDTO> getBanksByCountryCode(String countryCode) {
